requestの構造体をModel/APIRequestへ移動
