+++ QiitaClientApp/QiitaClientApp/Utility/QiitaAccessToken.swift
@@ -52,11 +52,3 @@ struct QiitaAccessToken {
             }
     }
 }
-
-struct AccessTokenDeactivateRequest: QiitaAPIRequest {
-    var parameters: Parameters { [:] }
-    
-    var path: String { "/access_tokens/\(accessToken)" }
-    var accessToken: String
-    let method = HTTPMethod.delete
-}
