+++ QiitaClientApp/QiitaClientApp/View/FeedViewController.swift
@@ -8,27 +8,6 @@
 import Alamofire
 import UIKit
 
-struct ArticlesGetRequest: QiitaAPIRequest {
-    public typealias Response = [QiitaArticle]
-    var path: String { "/items" }
-    var perPage: Int { 30 }
-    let method = HTTPMethod.get
-
-    let page: Int
-    let query: String?
-
-    var parameters: Parameters {
-        var parameters = [
-            "per_page": String(perPage),
-            "page": String(page)
-        ]
-        if let query = query {
-            parameters.updateValue(query, forKey: "query")
-        }
-        return parameters
-    }
-}
-
 final class FeedViewController: BaseArticlesViewController {
     var searchController: UISearchController!
     override func viewDidLoad() {
