+++ QiitaClientApp/QiitaClientApp/View/TagsViewController.swift
@@ -8,24 +8,6 @@
 import Alamofire
 import UIKit
 
-struct TagsGetRequest: QiitaAPIRequest {
-    public typealias Response = [QiitaTag]
-    var path: String { "/tags" }
-    var perPage: Int { 100 }
-    var sort: String { "count" }
-    let method = HTTPMethod.get
-
-    let page: Int
-
-    var parameters: Parameters {
-        [
-            "per_page": String(perPage),
-            "page": String(page),
-            "sort": sort
-        ]
-    }
-}
-
 struct CollectionSize {
     static let margin: CGFloat = 16
     static let cellWidth: CGFloat = 162
