+++ QiitaClientApp/QiitaClientApp/View/UserPageViewController.swift
@@ -8,29 +8,6 @@
 import Alamofire
 import UIKit
 
-struct AuthUserGetRequest: QiitaAPIRequest {
-    public typealias Response = QiitaUser
-    var path: String { "/authenticated_user" }
-    let method = HTTPMethod.get
-    var parameters: Parameters { [:] }
-}
-
-struct AuthUserArticlesGetRequest: QiitaAPIRequest {
-    public typealias Response = [QiitaArticle]
-    var path: String { "/authenticated_user/items" }
-    var perPage: Int { 30 }
-    let method = HTTPMethod.get
-
-    let page: Int
-
-    var parameters: Parameters {
-        [
-            "per_page": String(perPage),
-            "page": String(page)
-        ]
-    }
-}
-
 protocol UserPageViewControllerProtocol {
     func presentFollowViewController(followMode mode: FollowMode)
 }
